fix(agw): Enable opt-in based logging to Sentry
Makes it possible to select only a subset of log messages to be
sent to Sentry instead of everything above a certain log level.
In Python services there are now three possible configuration
settings for Sentry:
* send all errors
* send only selected errors (opt-in)
* disabled

Also adds a configurable decorator to send all uncaught errors
to Sentry, which is useful in the opt-in case. It is used for
all gRPC endpoints of mobilityd.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>